Formatter: LO-SRP
Extract out the responsibility of Formatting text for display (e.g. adding decorations) in to a separate class named Formatter from TextUi
