+++ src/seedu/addressbook/ui/Formatter.java
@@ -1,0 +1,76 @@
+package seedu.addressbook.ui;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import seedu.addressbook.data.person.ReadOnlyPerson;
+
+/**
+ * Text UI of the application.
+ */
+public class Formatter {
+
+    /**
+     * A decorative prefix added to the beginning of lines printed by AddressBook
+     */
+    private static final String LINE_PREFIX = "|| ";
+
+    /**
+     * A platform independent line separator.
+     */
+    private static final String LS = System.lineSeparator();
+
+    /**
+     * Format of indexed list item
+     */
+    private static final String MESSAGE_INDEXED_LIST_ITEM = "\t%1$d. %2$s";
+
+    /**
+     * Offset required to convert between 1-indexing and 0-indexing.
+     */
+    public static final int DISPLAYED_INDEX_OFFSET = 1;
+
+    public Formatter() { }
+
+    public static String appendPrefixTo(String message) {
+        return LINE_PREFIX + message;
+    }
+
+    public static String commandEntered(String command) {
+        return "[Command entered:" + command + "]";
+    }
+
+    public static String formatMessage(String message) {
+        return LINE_PREFIX + message.replace("\n", LS + LINE_PREFIX);
+    }
+
+    public static List<String> formatPersons(List<? extends ReadOnlyPerson> persons) {
+        final List<String> formattedPersons = new ArrayList<>();
+        for (ReadOnlyPerson person : persons) {
+            formattedPersons.add(person.getAsTextHidePrivate());
+        }
+        return formattedPersons;
+    }
+
+    /**
+     * Formats a list of strings as a viewable indexed list.
+     */
+    public static String getIndexedListForViewing(List<String> listItems) {
+        final StringBuilder formatted = new StringBuilder();
+        int displayIndex = DISPLAYED_INDEX_OFFSET;
+        for (String listItem : listItems) {
+            formatted.append(Formatter.getIndexedListItem(displayIndex, listItem)).append("\n");
+            displayIndex++;
+        }
+        return formatted.toString();
+    }
+
+    /**
+     * Formats a string as a viewable indexed list item.
+     *
+     * @param visibleIndex visible index for this listing
+     */
+    private static String getIndexedListItem(int visibleIndex, String listItem) {
+        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
+    }
+}
